pyproject.toml
[project]
name = "senuhas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "plotly>=6.1.1",
    "streamlit>=0.1.0",
]
